improve delete nic
up to this patch, if the API was returning an empty array on ReadNics
after a DeleteNic, the delete_nic operation on terraform was waiting for
the end of the timeout to validate the delete.
Not anymore.

Note that this bug did happen on Ricochet, but was not hapenning on the API
because the API was returning a nic either in deleting or deleted state.
But it was still a bug nevertheless.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>